+++ custom_components/ithodaalderop/sensors/base.py
@@ -1,7 +1,6 @@
 """Sensor base class."""
 
 import json
-from typing import Any
 
 from homeassistant.components import mqtt
 from homeassistant.components.binary_sensor import BinarySensorEntity
@@ -13,14 +12,9 @@ from homeassistant.helpers.device_registry import DeviceInfo
 from ..const import (
     ADDON_TYPES,
     CONF_ADDON_TYPE,
-    CONF_ADVANCED_CONFIG,
-    CONF_CUSTOM_BASETOPIC,
-    CONF_CUSTOM_DEVICE_NAME,
-    CONF_CUSTOM_ENTITY_PREFIX,
     CONF_NONCVE_MODEL,
     DOMAIN,
     MANUFACTURER,
-    MQTT_DEFAULT_BASETOPIC,
     NONCVE_DEVICES,
     UNITTYPE_ICONS,
 )
@@ -28,48 +22,7 @@ from ..definitions.base import (
     IthoBinarySensorEntityDescription,
     IthoSensorEntityDescription,
 )
-
-
-def get_default_mqtt_base_topic(config: dict[str, Any]) -> str:
-    """Get the default MQTT base topic."""
-    return MQTT_DEFAULT_BASETOPIC[config[CONF_ADDON_TYPE]]
-
-
-def get_mqtt_base_topic(config: dict[str, Any]) -> str:
-    """Get the MQTT base topic."""
-    if config[CONF_ADVANCED_CONFIG]:
-        return config[CONF_CUSTOM_BASETOPIC]
-
-    return get_default_mqtt_base_topic(config)
-
-
-def get_device_model(config: dict[str, Any]) -> str:
-    """Get the device model."""
-    if config[CONF_ADDON_TYPE] == "noncve":
-        return NONCVE_DEVICES[config[CONF_NONCVE_MODEL]]
-
-    return ADDON_TYPES[config[CONF_ADDON_TYPE]]
-
-
-def get_device_name(config: dict[str, Any]) -> str:
-    """Get the device name."""
-    if config[CONF_ADVANCED_CONFIG]:
-        return config[CONF_CUSTOM_DEVICE_NAME]
-
-    return get_device_model(config)
-
-
-def get_default_entity_prefix(config: dict[str, Any]) -> str:
-    """Get the default entity prefix."""
-    return f"itho_{ADDON_TYPES[config[CONF_ADDON_TYPE]].lower()}"
-
-
-def get_entity_prefix(config: dict[str, Any]) -> str:
-    """Get the entity prefix."""
-    if config[CONF_ADVANCED_CONFIG]:
-        return config[CONF_CUSTOM_ENTITY_PREFIX]
-
-    return get_default_entity_prefix(config)
+from ..vars import get_device_model, get_device_name, get_entity_prefix
 
 
 class IthoBaseSensor(SensorEntity):
